requirements.txt
pdf2image
opencv-python
numpy
pillow
anthropic
openai
qdrant-client
python-dotenv
jsonschema